Rename gradients.ts to colors.ts
